Fix project type selection
